Resolve PR comments, pt. 2
