+++ lib/espresso-runner.js
@@ -8,6 +8,7 @@ import { version } from '../../package.json'; // eslint-disable-line import/no-u
 import request from 'request-promise';
 import B from 'bluebird';
 import _ from 'lodash';
+import { copyRecursively } from './utils';
 
 
 const TEST_SERVER_PATH = path.resolve(__dirname, '..', '..', 'espresso-server');
@@ -16,30 +17,6 @@ const TEST_APK_PKG = 'io.appium.espressoserver.test';
 const REQUIRED_PARAMS = ['adb', 'tmpDir', 'host', 'systemPort', 'devicePort', 'appPackage', 'forceEspressoRebuild', 'espressoBuildConfig'];
 const ESPRESSO_SERVER_LAUNCH_TIMEOUT = 30000;
 const TARGET_PACKAGE_CONTAINER = '/data/local/tmp/espresso.apppackage';
-
-/**
- * Recursively copy all files
- * @param sourceBaseDir {string} directory to copy files from
- * @param targetBaseDir {string} directory to copy files to
- */
-async function copyRecursively (sourceBaseDir, targetBaseDir) {
-  await fs.walkDir(sourceBaseDir, true, async (itemPath, isDirectory) => {
-    const relativePath = path.relative(sourceBaseDir, itemPath);
-    const targetPath = path.resolve(targetBaseDir, relativePath);
-
-    const isGradleBuildDir = `${path.sep}${itemPath}${path.sep}`.includes(`${path.sep}build${path.sep}`);
-    if (isGradleBuildDir) {
-      return false;
-    }
-
-    if (!isDirectory) {
-      await fs.copyFile(itemPath, targetPath);
-    } else {
-      await mkdirp(targetPath);
-    }
-    return false;
-  });
-}
 
 class EspressoRunner {
   constructor (opts = {}) {
