+++ lib/server-builder.js
@@ -47,14 +47,13 @@ class ServerBuilder {
 
   getCommand () {
     let cmd = `.${path.sep}gradlew`;
-    let args = [];
-
-    for (const key of VERSION_KEYS) {
-      const serverVersion = this.server_versions[key];
-      if (key !== GRADLE_VERSION_KEY && serverVersion) {
-        args.push(`-P${key}=${serverVersion}`);
-      }
-    }
+    let args = VERSION_KEYS
+      .filter(key => key !== GRADLE_VERSION_KEY)
+      .map(key => {
+        const serverVersion = this.server_versions[key];
+        return serverVersion ? `-P${key}=${serverVersion}` : null;
+      })
+      .filter(serverVersion => serverVersion);
 
     args.push('assembleAndroidTest');
 
@@ -80,7 +79,7 @@ class ServerBuilder {
       .join('\n');
   }
 
-  runBuildProcess () {
+  async runBuildProcess () {
     const {cmd, args} = this.getCommand();
     log.debug(`Beginning build with command '${cmd} ${args.join(' ')}' ` +
       `in directory '${this.serverPath}'`);
@@ -118,7 +117,7 @@ class ServerBuilder {
 
     // wrap the start procedure in a promise so that we can catch, and report,
     // any startup errors that are thrown as events
-    return new B((resolve, reject) => {
+    return await new B((resolve, reject) => {
       gradlebuild.on('exit', (code, signal) => {
         log.error(`Gradle exited with code '${code}' and signal '${signal}'`);
