+++ lib/utils.js
@@ -1,4 +1,6 @@
+import { fs, mkdirp } from 'appium-support';
 import _ from 'lodash';
+import path from 'path';
 
 /**
  * https://android.googlesource.com/platform/frameworks/base/+/master/tools/aapt/Resource.cpp#755
@@ -22,4 +24,28 @@ function qualifyActivityName (activityName, packageName) {
   return `${packageName}${dotPos === 0 ? '' : '.'}${activityName}`;
 }
 
-export { qualifyActivityName };
+/**
+ * Recursively copy all files
+ * @param sourceBaseDir {string} directory to copy files from
+ * @param targetBaseDir {string} directory to copy files to
+ */
+async function copyRecursively (sourceBaseDir, targetBaseDir) {
+  await fs.walkDir(sourceBaseDir, true, async (itemPath, isDirectory) => {
+    const relativePath = path.relative(sourceBaseDir, itemPath);
+    const targetPath = path.resolve(targetBaseDir, relativePath);
+
+    const isGradleBuildDir = `${path.sep}${itemPath}${path.sep}`.includes(`${path.sep}build${path.sep}`);
+    if (isGradleBuildDir) {
+      return false;
+    }
+
+    if (!isDirectory) {
+      await fs.copyFile(itemPath, targetPath);
+    } else {
+      await mkdirp(targetPath);
+    }
+    return false;
+  });
+}
+
+export { qualifyActivityName, copyRecursively };
